refactor runs table
