+++ frontend/app/runs/page.tsx
@@ -1,12 +1,8 @@
 'use client';
 import OverviewContainer from '@/components/containers/OverviewContainer';
 import PageHeader from '@/components/headers/PageHeader';
-import { useAccount } from '@/components/providers/account-provider';
-import SkeletonTable from '@/components/skeleton/SkeletonTable';
-import { useGetJobRuns } from '@/libs/hooks/useGetJobRuns';
 import { ReactElement } from 'react';
-import { getColumns } from './components/JobRunsTable/columns';
-import { DataTable } from './components/JobRunsTable/data-table';
+import RunsTable from './components/RunsTable';
 
 export default function JobRuns(): ReactElement {
   return (
@@ -19,33 +15,7 @@ export default function JobRuns(): ReactElement {
       }
       containerClassName="runs-page"
     >
-      <JobRunsTable />
+      <RunsTable />
     </OverviewContainer>
-  );
-}
-
-interface TableProps {}
-
-function JobRunsTable(props: TableProps): ReactElement {
-  const {} = props;
-  const account = useAccount();
-  const { isLoading, data, mutate } = useGetJobRuns(account?.id ?? '');
-
-  if (isLoading) {
-    return <SkeletonTable />;
-  }
-
-  const runs = data?.jobRuns ?? [];
-
-  const columns = getColumns({
-    onDeleted() {
-      mutate();
-    },
-  });
-
-  return (
-    <div>
-      <DataTable columns={columns} data={runs} />
-    </div>
   );
 }
